+++ src/main/java/com/agh/surveys/service/poll/IPollService.java
@@ -9,9 +9,9 @@ public interface IPollService {
 
     List<Poll> findAll();
 
-    void deletePoll(Long id);
+    void deletePoll(Integer id);
 
-    Poll getPoll(Long id);
+    Poll getPoll(Integer id);
 
     void savePoll(Poll poll);
 }
